chore: move count total and filtered to abstract
